+++ src/sites/mobile-taro/vue/config/index.js
@@ -55,6 +55,10 @@ const config = {
         enable: true,
         config: {}
       },
+      pxtransform: {
+        enable: true,
+        config: { selectorBlackList: ['nut-'] }
+      },
       cssModules: {
         enable: false, // 默认为 false，如需使用 css modules 功能，则设为 true
         config: {
